# swagger_api_graph.yaml
swagger: "2.0"
info:
    version: "1.0.0"
    title: "ORJ Elysee"
host: "localhost:8080"
tags:
- name: "Armory"
  description: "Everything about Elysee's armory"
- name: "Intervention_group"
  description: "Access to available Intervention Group"
- name: "Mission"
  description: "Define Mission of your CRS, gouvern and conquer"
schemes:
- "http"
paths:
    /armory:
        get:
            tags:
            - "Armory"
            summary: "Get List of content of Armory"
            consumes:
            - "application/json"
            produces:
            - "application/json"
            - "application/xml"
            responses:
                200:
                    description: "Ok"
                    schema:
                        type: array
                        items:
                            $ref: "#/definitions/Armory"


    /mission:
        get:
            tags:
            - "Mission"
            summary: "Get List of content of Defined Mission"
            consumes:
            - "application/json"
            produces:
            - "application/json"
            responses:
                200:
                    description: "Ok"
                    schema:
                        type: array
                        items:
                            $ref: "#/definitions/Armory"

    /mission/{missionId}:
        get:
            tags:
            - "Mission"
            summary: "Get A mission by Id"
            produces:
            - "application/json"
            parameters:
            - name: "missionId"
              in: "path"
              description: "ID of Mission"
              required: true
              type: "string"
            responses:
                200:
                    description: "Ok"
                    schema:
                        $ref: "#/definitions/Mission"
                400:
                    description: "Invalid Id"
                    schema:
                        $ref: "#/definitions/ApiResponse"
        post:
            tags:
            - "Mission"
            summary: "Add a new mission, Sir ! Yes ! Sir !"
            produces:
            - "application/json"
            parameters:
            - name: "missionId"
              in: "path"
              description: "ID of Mission"
              required: true
              type: "string"
            - in: "body"
              name: "body"
              description: "Created user object"
              required: true
              schema:
                  $ref: "#/definitions/Mission"
            responses:
                201:
                    description: Mission Created
                    schema:
                        $ref: "#/definitions/Mission"
                400:
                    description: "Invalid Id"
                    schema:
                        $ref: "#/definitions/ApiResponse"
        delete:
            tags:
            - "Mission"
            summary: "Abort mission"
            parameters:
            - name: "missionId"
              in: "path"
              description: "ID of Mission"
              required: true
              type: "string"
            responses:
                204:
                    description: Ok, No Content
                400:
                    description: "Invalid Id"
                    schema:
                        $ref: "#/definitions/ApiResponse"
        put:
            tags:
            - "Mission"
            summary: "Update Mission"
            produces:
            - "application/json"
            parameters:
            - name: "missionId"
              in: "path"
              description: "ID of Mission"
              required: true
              type: "string"
            - in: "body"
              name: "body"
              description: "Created user object"
              required: true
              schema:
                  $ref: "#/definitions/Mission"
            responses:
                200:
                    description: OK
                    schema:
                        $ref: "#/definitions/Mission"
                400:
                    description: "Invalid Id"
                    schema:
                        $ref: "#/definitions/ApiResponse"


definitions:
    Mission:
        type: "object"
        properties:


    Armory:
        type: "object"
        properties:
            id:
                type: "string"
            cost:
                type: "integer"
            stock:
                type: "integer"
            name:
                type: "string"
        xml:
            name: "Armory"
    ApiResponse:
        type: "object"
        properties:
            code:
                type: "integer"
                format: "int32"
            type:
                type: "string"
            message:
                type: "string"
externalDocs:
    description: "Find out more about Swagger"
    url: "http://swagger.io"